## 4.5.1.0
    * Bug Fixes
      * Ellipse Vertex radial should not have a delete option
      * Adding height to an open polyline closes the shape
      * Wrong text entry field opens initially when selecting to change a shape's height
      * Fix crash `NullPointerException` `EncryptedEditText`
      * WFS Feature Styling not working for Discovered Features
      * Verify Logic Bombs (Billion Laughs and ZipBomb) don't crash ATAK
      * Crash `NullPointerException` `GLMapMenuButtonWidget`
      * GeoPDF datapackage crashes when loading.
      * Fix performance issue with Bread crumbs
    * Feature Additions
      * Allow for Inner Signing of Plugins to allow for third party developers make use of AAB submission to the play store.
      * Augment Feature Editing
      * Additional Drawing and Line Capabilities to include Dotted, Dashed and Solid lines. Ability to select different fill and line colors for closed shapes.
      * Left Right Switch for the overflow button
      * Allow for sorting of IP's
      * Optional lower bar for the Self Widget Information

+++ atak/ATAK/app/src/main/java/com/atakmap/spatial/file/MvtMapOverlay.java
@@ -58,7 +58,7 @@ public final class MvtMapOverlay extends FeatureDataStoreMapOverlay {
             for (HierarchyListItem i : items) {
                 try {
                     final FeatureSet fs = Utils.getFeatureSet(spatialDb,
-                            ((Long) i.getUserObject()).longValue());
+                            (Long) i.getUserObject());
                     if (fs != null) {
                         final File f = Utils.getSourceFile(spatialDb, fs);
                         List<HierarchyListItem> mapped = m.get(f);
